Bunch of fixes
