+++ vj4/ui/components/ide/ide-constants.js
@@ -1,0 +1,113 @@
+export default {
+  supportLangTypes: {
+    c: 'C',
+    cc: 'C++',
+    pas: 'Pascal',
+    java: 'Java',
+    cs: 'C#',
+    py: 'Python 2.x',
+    py3: 'Python 3',
+    js: 'JavaScript',
+    php: 'PHP',
+    ruby: 'Ruby',
+  },
+  mapLangToMode: {
+    c: 'text/x-csrc',
+    cc: 'text/x-c++src',
+    pas: 'text/x-pascal',
+    java: 'text/x-java',
+    cs: 'text/x-csharp',
+    py: {
+      name: 'text/x-cython',
+      version: 2,
+      singleLineStringErrors: false,
+    },
+    py3: {
+      name: 'python',
+      version: 3,
+      singleLineStringErrors: false,
+    },
+    js: 'text/javascript',
+    php: 'text/x-php',
+    ruby: 'text/x-ruby',
+  },
+  codeToStatusMap: {
+    0: 'STATUS_WAITING',
+    1: 'STATUS_ACCEPTED',
+    2: 'STATUS_WRONG_ANSWER',
+    3: 'STATUS_TIME_LIMIT_EXCEEDED',
+    4: 'STATUS_MEMORY_LIMIT_EXCEEDED',
+    5: 'STATUS_OUTPUT_LIMIT_EXCEEDED',
+    6: 'STATUS_RUNTIME_ERROR',
+    7: 'STATUS_COMPILE_ERROR',
+    8: 'STATUS_SYSTEM_ERROR',
+    9: 'STATUS_CANCELED',
+    10: 'STATUS_ETC',
+    20: 'STATUS_JUDGING',
+    21: 'STATUS_COMPILING',
+    30: 'STATUS_IGNORED',
+  },
+  statusToClassNameMap: {
+    STATUS_WAITING: 'result--pending',
+    STATUS_ACCEPTED: 'result--pass',
+    STATUS_WRONG_ANSWER: 'result--fail',
+    STATUS_TIME_LIMIT_EXCEEDED: 'result--fail',
+    STATUS_MEMORY_LIMIT_EXCEEDED: 'result--fail',
+    STATUS_OUTPUT_LIMIT_EXCEEDED: 'result--fail',
+    STATUS_RUNTIME_ERROR: 'result--fail',
+    STATUS_COMPILE_ERROR: 'result--fail',
+    STATUS_SYSTEM_ERROR: 'result--fail',
+    STATUS_CANCELED: 'result--ignore',
+    STATUS_ETC: 'result--fail',
+    STATUS_JUDGING: 'result--progress',
+    STATUS_COMPILING: 'result--progress',
+    STATUS_IGNORED: 'result--ignore',
+  },
+  statusToNotation: {
+    STATUS_WAITING: 'PENDING',
+    STATUS_ACCEPTED: 'AC',
+    STATUS_WRONG_ANSWER: 'WA',
+    STATUS_TIME_LIMIT_EXCEEDED: 'TLE',
+    STATUS_MEMORY_LIMIT_EXCEEDED: 'MLE',
+    STATUS_OUTPUT_LIMIT_EXCEEDED: 'OLE',
+    STATUS_RUNTIME_ERROR: 'RE',
+    STATUS_COMPILE_ERROR: 'CE',
+    STATUS_SYSTEM_ERROR: 'IE',
+    STATUS_CANCELED: 'CANCELED',
+    STATUS_ETC: 'IE',
+    STATUS_JUDGING: 'WAITING',
+    STATUS_COMPILING: 'WAITING',
+    STATUS_IGNORED: 'IGNORED',
+  },
+  statusToContentNameMap: {
+    STATUS_WAITING: 'Pending',
+    STATUS_ACCEPTED: 'Accepted',
+    STATUS_WRONG_ANSWER: 'Wrong Answer',
+    STATUS_TIME_LIMIT_EXCEEDED: 'Time Limit Exceeded',
+    STATUS_MEMORY_LIMIT_EXCEEDED: 'Memory Limit Exceeded',
+    STATUS_OUTPUT_LIMIT_EXCEEDED: 'Output Limit Exceeded',
+    STATUS_RUNTIME_ERROR: 'Runtime Error',
+    STATUS_COMPILE_ERROR: 'Compile Error',
+    STATUS_SYSTEM_ERROR: 'System Error',
+    STATUS_CANCELED: 'Canceled',
+    STATUS_ETC: 'Unknown Error',
+    STATUS_JUDGING: 'Judging',
+    STATUS_COMPILING: 'Compiling',
+    STATUS_IGNORED: 'Ignored',
+  },
+  statusToWillShowDetail: {
+    STATUS_WAITING: false,
+    STATUS_ACCEPTED: true,
+    STATUS_WRONG_ANSWER: true,
+    STATUS_TIME_LIMIT_EXCEEDED: true,
+    STATUS_MEMORY_LIMIT_EXCEEDED: true,
+    STATUS_RUNTIME_ERROR: true,
+    STATUS_COMPILE_ERROR: false,
+    STATUS_SYSTEM_ERROR: false,
+    STATUS_CANCELED: false,
+    STATUS_ETC: false,
+    STATUS_JUDGING: false,
+    STATUS_COMPILING: false,
+    STATUS_IGNORED: false,
+  },
+};
